pyproject.toml
[project]
name = "commodity-sentiment-monitor"
version = "0.1.0"
description = "Real-time commodity sentiment analysis from audio streams"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.13.0",
    "pydantic>=2.10.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.0",
    "yt-dlp>=2024.0.0",
    "openai>=2.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]

[project.scripts]
csm = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "api: integration tests that call live LLM APIs",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "All"
